move affine_grid code from modules to functional
