+++ python/oneflow/nn/modules/affine_grid.py
@@ -15,58 +15,17 @@ limitations under the License.
 """
 from typing import List
 
-import oneflow as flow
+from oneflow.nn.modules.module import Module
+import oneflow.nn.functional as F
 
 
-def affine_grid(theta, size: List[int], align_corners: bool = False):
-    """The interface is consistent with PyTorch.    
-    The documentation is referenced from: 
-    https://pytorch.org/docs/1.10/generated/torch.nn.functional.affine_grid.html.
-
-    Generates a 2D or 3D flow field (sampling grid), given a batch of
-    affine matrices :attr:`theta`.
-
-    .. note::
-        This function is often used in conjunction with :func:`grid_sample`
-        to build `Spatial Transformer Networks`_ .
-
-    Args:
-        theta (Tensor): input batch of affine matrices with shape
-            (:math:`N, 2, 3`) for 2D or
-            (:math:`N, 3, 4`) for 3D
-        size (oneflow.Size): the target output image size.
-            (:math:`N, C, H, W` for 2D or
-            :math:`N, C, D, H, W` for 3D)
-            Example: oneflow.Size((32, 3, 24, 24))
-        align_corners (bool): if ``True``, consider ``-1`` and ``1``
-            to refer to the centers of the corner pixels rather than the image corners.
-            Refer to :func:`grid_sample` for a more complete description.
-            A grid generated by :func:`affine_grid` should be passed to :func:`grid_sample`
-            with the same setting for this option.
-            Default: ``False``
-
-    Returns:
-        output (Tensor): output Tensor of size (:math:`N, H, W, 2`)
-
-    .. _`Spatial Transformer Networks`:
-        https://arxiv.org/abs/1506.02025
-
-    Examples::
-
-        >>> import oneflow as flow
-        >>> import numpy as np
-        >>> input = flow.tensor(np.arange(1., 7).reshape((1, 2, 3)), dtype=flow.float32)
-        >>> output = flow.nn.functional.affine_grid(input, flow.Size([1, 1, 2, 2]), align_corners=True)
-        >>> output
-        tensor([[[[ 0., -3.],
-                  [ 2.,  5.]],
-        <BLANKLINE>
-                 [[ 4.,  7.],
-                  [ 6., 15.]]]], dtype=oneflow.float32)
-    """
-    y = flow._C.affine_grid(theta, size=size, align_corners=align_corners)
-    return y
-
+class AffineGrid(Module):
+    def __init__(self, size: List[int], align_corners: bool = False):
+        super().__init__()
+        self.size = size
+        self.align_corners = align_corners
+    def forward(self, x):
+        return F.affine_grid(x, self.size, self.align_corners)
 
 if __name__ == "__main__":
     import doctest
